2lab/project/requirements.txt
alembic==1.13.3
bcrypt==4.0.1
celery==5.4.0
fastapi==0.115.0
numpy==2.1.1
opencv-python==4.10.0.84
passlib[bcrypt]==1.7.4
pillow==10.4.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
pyjwt
python-jose[cryptography]==3.3.0
python-multipart
redis==5.0.8
scikit-image
sqlalchemy==2.0.35
uvicorn==0.30.6
wsproto